backend2/requirements.txt
kafka-python==2.0.2
redis==4.1.0
